+++ massgen/v2/backends/__init__.py
@@ -1,0 +1,17 @@
+"""
+Backend implementations for different LLM providers.
+"""
+
+from .base import AgentBackend, TokenUsage
+from .chat_completions import ChatCompletionsBackend  
+from .openai import OpenAIResponseBackend
+from .factory import create_backend, get_provider_from_model
+
+__all__ = [
+    'AgentBackend',
+    'TokenUsage', 
+    'ChatCompletionsBackend',
+    'OpenAIResponseBackend',
+    'create_backend',
+    'get_provider_from_model'
+]
