+++ massgen/v2/backends/openai.py
@@ -1,177 +1,12 @@
 """
-AgentBackend Architecture - Issue #18: Implement AgentBackend Architecture
-
-This module implements a comprehensive AgentBackend architecture with async streaming,
-token tracking, and multi-provider support.
+OpenAI backend implementation.
 """
 
-from abc import ABC, abstractmethod
-from typing import Dict, List, Optional, Any, AsyncGenerator, Callable
-from dataclasses import dataclass
-import asyncio
-import time
-from .chat_agent import StreamChunk
-
-
-@dataclass
-class TokenUsage:
-    """Token usage and cost tracking."""
-    input_tokens: int = 0
-    output_tokens: int = 0
-    estimated_cost: float = 0.0
-    model: Optional[str] = None
-    provider: Optional[str] = None
-    timestamp: Optional[float] = None
-
-    def __post_init__(self):
-        """Set timestamp if not provided."""
-        if self.timestamp is None:
-            self.timestamp = time.time()
-    
-    def add_usage(self, input_tokens: int, output_tokens: int, cost: float = 0.0) -> None:
-        """Add token usage to the current totals."""
-        self.input_tokens += input_tokens
-        self.output_tokens += output_tokens
-        self.estimated_cost += cost
-    
-    def get_total_tokens(self) -> int:
-        """Get total tokens used."""
-        return self.input_tokens + self.output_tokens
-
-
-class AgentBackend(ABC):
-    """Abstract base class for agent LLM providers."""
-    
-    def __init__(self, **kwargs):
-        """
-        Initialize the AgentBackend.
-        
-        Args:
-            **kwargs: Configuration parameters including api_key, model, etc.
-        """
-        self.config = kwargs
-        self.token_usage = TokenUsage()
-    
-    @abstractmethod
-    async def stream_with_tools(self, 
-                               messages: List[Dict[str, Any]], 
-                               tools: Optional[List[Dict[str, Any]]] = None, 
-                               **kwargs) -> AsyncGenerator[StreamChunk, None]:
-        """
-        Stream response with tool calling support.
-        
-        Args:
-            messages: List of conversation messages
-            tools: Optional list of tools available to the agent
-            **kwargs: Additional parameters for the specific backend
-            
-        Yields:
-            StreamChunk: Standardized streaming response chunks
-        """
-        pass
-    
-    @abstractmethod
-    def get_provider_name(self) -> str:
-        """
-        Get provider name.
-        
-        Returns:
-            str: Name of the provider (e.g., "anthropic", "google", "openai", "xai")
-        """
-        pass
-    
-    @abstractmethod
-    def calculate_cost(self, input_tokens: int, output_tokens: int, model: str) -> float:
-        """
-        Calculate cost for token usage.
-        
-        Args:
-            input_tokens: Number of input tokens
-            output_tokens: Number of output tokens
-            model: Model name used
-            
-        Returns:
-            float: Estimated cost in USD
-        """
-        pass
-    
-    def get_token_usage(self) -> TokenUsage:
-        """
-        Get current token usage statistics.
-        
-        Returns:
-            TokenUsage: Current usage statistics
-        """
-        return self.token_usage
-    
-    def reset_token_usage(self) -> None:
-        """Reset token usage statistics."""
-        self.token_usage = TokenUsage()
-    
-    def get_status(self) -> Dict[str, Any]:
-        """
-        Get backend status information.
-        
-        Returns:
-            Dict containing backend status
-        """
-        return {
-            "provider": self.get_provider_name(),
-            "token_usage": {
-                "input_tokens": self.token_usage.input_tokens,
-                "output_tokens": self.token_usage.output_tokens,
-                "total_tokens": self.token_usage.get_total_tokens(),
-                "estimated_cost": self.token_usage.estimated_cost
-            },
-            "config": self.config
-        }
-
-
-class ChatCompletionsBackend(AgentBackend):
-    """Base class for standard chat completions APIs (OpenAI-compatible)."""
-    
-    def __init__(self, **kwargs):
-        """
-        Initialize ChatCompletionsBackend.
-        
-        Args:
-            **kwargs: Configuration including model, api_key, base_url, etc.
-        """
-        super().__init__(**kwargs)
-    
-    def format_tools_for_api(self, tools: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
-        """
-        Format tools for the specific API format.
-        
-        Args:
-            tools: List of tool definitions
-            
-        Returns:
-            List of formatted tools
-        """
-        # Default implementation - can be overridden by specific backends
-        formatted_tools = []
-        for tool in tools:
-            if isinstance(tool, dict):
-                formatted_tools.append(tool)
-            else:
-                # TODO: Handle other tool formats as needed
-                formatted_tools.append(tool)
-        return formatted_tools
-    
-    def estimate_tokens(self, text: str) -> int:
-        """
-        Estimate token count for text.
-        
-        Args:
-            text: Text to estimate tokens for
-            
-        Returns:
-            int: Estimated token count
-        """
-        # Simple estimation: ~4 characters per token
-        # TODO: Use a more accurate tokenizer if available
-        return len(text) // 4
+import os
+import warnings
+from typing import Dict, List, Optional, Any, AsyncGenerator
+from .base import AgentBackend
+from ..chat_agent import StreamChunk
 
 
 class OpenAIResponseBackend(AgentBackend):
@@ -190,7 +25,6 @@ class OpenAIResponseBackend(AgentBackend):
         # Import OpenAI client here to avoid import issues
         try:
             from openai import AsyncOpenAI
-            import os
             
             # Get API key from config or environment
             api_key_val = self.config.get("api_key") or os.getenv("OPENAI_API_KEY")
@@ -378,9 +212,6 @@ class OpenAIResponseBackend(AgentBackend):
             return input_cost + output_cost
         else:
             # Configurable fallback pricing for unknown models
-            import os
-            import warnings
-            
             fallback_input_rate = float(os.getenv("FALLBACK_INPUT_RATE", "0.0025"))
             fallback_output_rate = float(os.getenv("FALLBACK_OUTPUT_RATE", "0.01"))
             
@@ -393,48 +224,3 @@ class OpenAIResponseBackend(AgentBackend):
             )
             
             return (input_tokens / 1000) * fallback_input_rate + (output_tokens / 1000) * fallback_output_rate
-
-
-# Factory function for creating backends
-def create_backend(provider: str, model: str, **kwargs) -> AgentBackend:
-    """
-    Create an agent backend based on provider.
-    
-    Args:
-        provider: Provider name ("anthropic", "google", "openai", "xai", etc.)
-        model: Model name
-        **kwargs: Additional configuration
-        
-    Returns:
-        AgentBackend: Initialized backend instance
-    """
-    if provider.lower() == "openai":
-        return OpenAIResponseBackend(model=model, **kwargs)
-    else:
-        raise ValueError(f"Unsupported provider: {provider}")
-
-
-# Utility function to detect provider from model name
-def get_provider_from_model(model: str) -> str:
-    """
-    Detect provider from model name.
-    
-    Args:
-        model: Model name
-        
-    Returns:
-        str: Provider name
-    """
-    model_lower = model.lower()
-    
-    if any(prefix in model_lower for prefix in ["claude"]):
-        return "anthropic"
-    elif any(prefix in model_lower for prefix in ["gemini"]):
-        return "google"
-    elif any(prefix in model_lower for prefix in ["gpt", "o1", "o3"]):
-        return "openai"
-    elif any(prefix in model_lower for prefix in ["grok"]):
-        return "xai"
-    else:
-        # Default to OpenAI for unknown models
-        return "openai"
